# autoassist/apps/api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoAssist+ API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local (Docker Compose)" },
    { "url": "http://localhost:3001", "description": "Local (node dev)" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            },
            "required": ["code"]
          }
        },
        "required": ["error"]
      },
      "TokenPair": {
        "type": "object",
        "properties": {
          "access": { "type": "string", "description": "JWT access token" },
          "refresh": { "type": "string", "description": "JWT refresh token" }
        },
        "required": ["access", "refresh"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 }
        },
        "required": ["email", "password"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "name": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refresh": { "type": "string", "minLength": 10 }
        },
        "required": ["refresh"]
      },
...
      "Attachment": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "type": { "type": "string", "enum": ["DOCUMENT","PHOTO","VIDEO","AUDIO"] },
          "status": { "type": "string", "enum": ["pending","ready","removed"] },
          "filename": { "type": "string" },
          "objectKey": { "type": "string" },
          "contentType": { "type": "string" },
          "size": { "type": "integer" }
        }
      },
      "PresignUploadRequest": {
        "type": "object",
        "properties": {
          "orderId": { "type": "integer" },
          "fileName": { "type": "string" },
          "contentType": { "type": "string", "enum": ["image/jpeg","image/png","image/webp","image/heic","video/mp4","application/pdf"] },
          "size": { "type": "integer", "maximum": 26214400 },
          "kind": { "type": "string", "enum": ["photo","video","doc"] },
          "meta": { "type": "object", "additionalProperties": true }
        },
        "required": ["orderId","fileName","contentType","size"]
      },
      "PresignUploadResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "attachmentId": { "type": "integer" },
          "putUrl": { "type": "string" },
          "objectKey": { "type": "string" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } } },
          "409": { "description": "Email taken", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh tokens",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } } },
          "401": { "description": "Invalid refresh token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/attachments/presign-upload": {
      "post": {
        "summary": "Presign S3/MinIO upload and create attachment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PresignUploadRequest" },
              "examples": {
                "image": {
                  "summary": "JPEG photo upload",
                  "value": {
                    "orderId": 142,
                    "fileName": "damage-front.jpg",
                    "contentType": "image/jpeg",
                    "size": 3482304,
                    "kind": "photo",
                    "meta": { "angle": "front-left" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PresignUploadResponse" },
                "examples": {
                  "ok": {
                    "value": {
                      "id": 987,
                      "attachmentId": 654,
                      "putUrl": "http://localhost:9000/autoassist/uploads/damage-front.jpg?X-Amz-Algorithm=AWS4-HMAC-SHA256...",
                      "objectKey": "uploads/2025/10/22/142/damage-front.jpg"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/attachments/{id}/complete": {
      "post": {
        "summary": "Mark upload complete (generate preview async)",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "409": { "description": "Not ready/invalid" } }
      }
    },
    "/api/attachments/{id}/presign": {
      "get": {
        "summary": "Presign download URL",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "url": { "type": "string" } }, "required": ["url"] },
                "examples": { "ok": { "value": { "url": "http://localhost:9000/autoassist/uploads/2025/10/22/142/damage-front.jpg?X-Amz-Algorithm=AWS4-HMAC-SHA256..." } } }
              }
            }
          }
        }
      }
    },
    "/api/attachments/order/{orderId}": {
      "get": {
        "summary": "List attachments by order",
        "parameters": [{ "in": "path", "name": "orderId", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } } } } } }
      }
    },
    "/api/attachments/{id}": {
      "delete": {
        "summary": "Soft-delete attachment and enqueue cleanup",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/notifications/test": {
      "post": {
        "summary": "Send test notification (admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "userId": { "type": "integer" }, "type": { "type": "string" }, "title": { "type": "string" }, "message": { "type": "string" }, "priority": { "type": "string" }, "channels": { "type": "array", "items": { "type": "string" } } } }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" } }
      }
    },
...
  }
}